Extract tus utils
